+++ packages/stream_chat_flutter/lib/src/components/avatar/stream_user_avatar.dart
@@ -135,7 +135,8 @@ class StreamUserAvatar extends StatelessWidget {
     .xs || .sm => StreamOnlineIndicatorSize.sm,
     .md => StreamOnlineIndicatorSize.md,
     .lg => StreamOnlineIndicatorSize.lg,
-    .xl || .xxl => StreamOnlineIndicatorSize.xl,
+    .xl => StreamOnlineIndicatorSize.xl,
+    .xxl => StreamOnlineIndicatorSize.xxl,
   };
 }