+++ packages/stream_chat_flutter/lib/src/reactions/detail/reaction_detail_sheet.dart
@@ -178,8 +178,10 @@ class _ReactionDetailSheetState extends State<ReactionDetailSheet> {
           leading: StreamEmojiChip.addEmoji(
             onPressed: () async {
               final selectedReactions = ownReactionsMap.keys.toSet();
+              final supportedEmojis = resolver.supportedReactions.map((type) => streamSupportedEmojis[type]).nonNulls;
               final emoji = await StreamEmojiPickerSheet.show(
                 context: context,
+                emojis: supportedEmojis,
                 selectedReactions: selectedReactions,
               );